+++ ncpp/include/Colors.h
@@ -25,6 +25,8 @@ COPYRIGHT:	University Corporation for Atmospheric Research, 1997-2000
 #include <X11/Xutil.h>
 #include <X11/Xos.h>
 
+#include <stdint.h>
+
 #ifdef PNG
 #include <png.h>
 #endif
@@ -39,9 +41,9 @@ public:
   void	ResetColors()		{ colorIndex = 0; resetColors(); }
   void	SetColorNames(char str[]);
 
-  unsigned long	GetColor(int indx);
-  unsigned long	NextColor();
-  unsigned long	CurrentColor();
+  uint32_t	GetColor(int indx);
+  uint32_t	NextColor();
+  uint32_t	CurrentColor();
 
   float	*GetColorPS(int indx);
   float	*NextColorPS();
@@ -49,7 +51,7 @@ public:
 
 #ifdef PNG
   void	SavePNG(const char outFile[], XImage *image);
-  unsigned short	*GetColorRGB_X(int idx);
+  uint16_t	*GetColorRGB_X(int idx);
 #endif
 
 
@@ -58,7 +60,7 @@ private:
 
 #ifdef PNG
   void	checkByteSwap(XImage *image);
-  int	getColorIndex(unsigned long pixel);
+  int	getColorIndex(uint32_t pixel);
 #endif
 
   Widget _w;
